# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Films API - Documentation Swagger",
        "description": "Documentation complète du TP1 Films API",
        "version": "1.0.0"
    },
    "paths": {
        "/api/critics/{id}": {
            "delete": {
                "tags": [
                    "Critics"
                ],
                "summary": "Supprimer une critique",
                "operationId": "4c7bbd907e57760bbbfc560c1ba8e1ba",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Critique supprimée"
                    },
                    "404": {
                        "description": "Critique introuvable"
                    }
                }
            }
        },
        "/api/films": {
            "get": {
                "tags": [
                    "Films"
                ],
                "summary": "Liste paginée des films",
                "operationId": "271e474415d2f29bf3c163247c5f4639",
                "responses": {
                    "200": {
                        "description": "Liste paginée des films"
                    }
                }
            }
        },
        "/api/films/{id}/actors": {
            "get": {
                "tags": [
                    "Films"
                ],
                "summary": "Acteurs d’un film",
                "operationId": "e620e233489a783ac6cbd7b93c70fa97",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID du film",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Liste des acteurs"
                    },
                    "404": {
                        "description": "Film introuvable"
                    }
                }
            }
        },
        "/api/films/{id}/critics": {
            "get": {
                "tags": [
                    "Films"
                ],
                "summary": "Critiques d’un film",
                "operationId": "a47883c5b30fbdcc397806a9941440a0",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID du film",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Liste des critiques"
                    },
                    "404": {
                        "description": "Film introuvable"
                    }
                }
            }
        },
        "/api/films/{id}/average-score": {
            "get": {
                "tags": [
                    "Films"
                ],
                "summary": "Recevoir la moyenne des critiques d’un film",
                "operationId": "fcd693de18daad26767293c2f5a2927e",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID du film pour lequel on veut la moyenne",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Moyenne des critiques",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "film_id": {
                                            "type": "integer"
                                        },
                                        "average_score": {
                                            "type": "number",
                                            "format": "float"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Film introuvable"
                    }
                }
            }
        },
        "/api/films/search": {
            "get": {
                "tags": [
                    "Films"
                ],
                "summary": "Rechercher des films",
                "operationId": "ddbc3843cfe184eb1429a6ad3d501fbb",
                "parameters": [
                    {
                        "name": "keyword",
                        "in": "query",
                        "description": "Mot-clé dans le titre",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "rating",
                        "in": "query",
                        "description": "Filtrer par rating exact",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "minLength",
                        "in": "query",
                        "description": "Durée minimum",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "maxLength",
                        "in": "query",
                        "description": "Durée maximum",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Résultats de la recherche"
                    }
                }
            }
        },
        "/api/users": {
            "post": {
                "tags": [
                    "Users"
                ],
                "summary": "Créer un utilisateur",
                "operationId": "592819a0265360b2014512d6dbfaf0e7",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "email",
                                    "password",
                                    "language_id"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string"
                                    },
                                    "email": {
                                        "type": "string",
                                        "format": "email"
                                    },
                                    "password": {
                                        "type": "string"
                                    },
                                    "language_id": {
                                        "type": "integer"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Utilisateur créé"
                    },
                    "422": {
                        "description": "Erreur de validation"
                    }
                }
            }
        },
        "/api/users/{id}": {
            "put": {
                "tags": [
                    "Users"
                ],
                "summary": "Mettre à jour un utilisateur",
                "operationId": "b9091397c8b25f12c6adb74be6ce3a5a",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "email",
                                    "password",
                                    "language_id"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string"
                                    },
                                    "email": {
                                        "type": "string"
                                    },
                                    "password": {
                                        "type": "string"
                                    },
                                    "language_id": {
                                        "type": "integer"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Utilisateur mis à jour"
                    },
                    "404": {
                        "description": "Utilisateur introuvable"
                    },
                    "422": {
                        "description": "Erreur de validation"
                    }
                }
            }
        },
        "/api/users/{id}/language": {
            "get": {
                "tags": [
                    "Users"
                ],
                "summary": "Recevoir le langage préféré d’un utilisateur",
                "description": "Détermine le langage préféré d’un utilisateur basé sur les films qu'il a critiqués. En cas d’égalité, retourne celui qui apparaît en premier.",
                "operationId": "e4c55e70f04fde580331dd30dfb9bf90",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID de l'utilisateur",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Langage préféré détecté ou message si aucune critique",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "oneOf": [
                                        {
                                            "properties": {
                                                "id": {
                                                    "type": "integer"
                                                },
                                                "name": {
                                                    "type": "string"
                                                }
                                            },
                                            "type": "object"
                                        },
                                        {
                                            "properties": {
                                                "message": {
                                                    "type": "string"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    ]
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Utilisateur introuvable"
                    }
                }
            }
        }
    },
    "tags": [
        {
            "name": "Critics",
            "description": "Critics"
        },
        {
            "name": "Films",
            "description": "Films"
        },
        {
            "name": "Users",
            "description": "Users"
        }
    ]
}